table and sorting

+++ src/app/productlist/productlist.component.ts
@@ -6,6 +6,7 @@ import { Province } from '../province';
 import {MatSort, Sort} from '@angular/material/sort';
 import {MatTableDataSource} from '@angular/material/table';
 import {LiveAnnouncer} from '@angular/cdk/a11y'
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-productlist',
@@ -14,10 +15,25 @@ import {LiveAnnouncer} from '@angular/cdk/a11y'
 })
 export class ProductlistComponent implements OnInit {
   products: Array<Product> = [];
+  
+  page = 1;
+  pageSize = 4;
+  collectionSize = this.products.length;
+  prolist: Product[];
 
   constructor(private _route: Router, private _service: NgserviceService, private _liveAnnouncer: LiveAnnouncer) { }
+
+  user: string;
+
   ngOnInit(): void {
     this.getProducts();
+   // this.user = this.isUserLoggedIn('User');
+    this.refreshCountries();
+  }
+  refreshCountries() {
+    this.prolist = this.products
+      .map((country, i) => ({id: i + 1, ...country}))
+      .slice((this.page - 1) * this.pageSize, (this.page - 1) * this.pageSize + this.pageSize);
   }
 
   getProducts() {
@@ -50,13 +66,55 @@ export class ProductlistComponent implements OnInit {
   }
 
   deleteProduct(id: number) {
-    if (confirm('Are you sure ?'))
-  return this._service.deleteBdyIdFromRemote(id).subscribe(
-    success =>{
-      ("Product deleted succesfully");
-    },
-    error=> {console.log("Exception occured 2"); this.getProducts()}
-   )
-  }
+   Swal.fire({  
+    title: 'Are you sure want to remove?',  
+    text: 'You will not be able to recover this file!',  
+    icon: 'warning',  
+    showCancelButton: true,  
+    confirmButtonText: 'Yes, delete it!',  
+    cancelButtonText: 'No, keep it'  
+  }).then((result) => {  
+    if (result.value) {  
+      Swal.fire(  
+        'Deleted!',  
+        'Your imaginary file has been deleted.',  
+        'success',
+      ) 
+      return this._service.deleteBdyIdFromRemote(id).subscribe(
+        success =>{
+          ("Product deleted succesfully");
+        },
+        error=> {console.log("Exception occured 2"); this.getProducts()}
+       ) 
+    } else if (result.dismiss === Swal.DismissReason.cancel) {  
+      Swal.fire(  
+        'Cancelled',  
+        'Your imaginary file is safe :)',  
+        'error'  
+      )  
+    }  
+  }) 
+}
 
+  isUserLoggedIn(User: string): string {
+    User = localStorage.getItem(User);
+    if (User === null) {
+      Swal.fire({
+        icon: 'error',  
+        text: 'Something went wrong!', 
+        position: 'center',
+        title: 'Please Login!',
+        showConfirmButton: false,
+        timer: 2000,
+      });
+    this._route.navigate(['/login'])
+    }
+    return User;
+}
+
+logout() {
+  localStorage.clear();
+  console.log("logout success");
+  this._route.navigate(['/login'])
+}
 }
